[IMP] web: data_manager: make field description self-contained
The fields are represented as an Object whose keys are the field
names and values their description. By adding the name inside the
description, this commit makes it fully self-contained.

This commit also moves the processing of fields_views s.t. they
are stored processed in the cache, instead of processing them each
time they were requested.