[FIX] account: usability of entry sequences.
This revert commits 1258916a256c82aac1f96da1791fc75e22a48dd0 and 8c01877c6d7e33f42e00bf8227290267e093a7d8 that were introduced to solve OPW-693955.

The usability before those fixes was good, but the bug was appearing when we first create the journal without the boolean 'dedicated refund sequence' set, then we try to modify that va$

This time, the fix consist to create automatically the sequence on the write of the boolean if it's not existing yet, in the same fashion than on the create().