[FIX] stock: multiple global push rules with same origin
When you have multiplle global push rules with the same source location,
a traceback is raised because the '_apply' method of
"stock.location.path" is called with more than one record, and doesn't
handle multiple record.

To fix this, we limit the number of records returned when a global rule
is searched to 1 (as for other 'search' of push rules in this file).

This bug has been introduced during migration in rev: https://github.com/odoo/odoo/commit/e308d229430e02e18922117b7b3ef43fd5de2864

opw-725874