[FIX] web: debounce field changes for input fields
Before this commit, every field changes were sent upstream and triggered
a notifyChange, potentially causing multiple (onchanges) RPCS in rapid
succession.

It is useful to have a feedback of the new values of the fields even
without needing to lose the focus, but it should not be done for each
keystroke.

With this commit, we debounce the onInput handler, so the method
setValue is in practice debounced as well.  We need to add a handler on
the focusout event to make sure that if the user changes input, the
setValue is immediately called.

Note that we had to add support for that in the test utils environment
(and not only in the createView method: field widgets can be
instantiated in other environment, such as the reconciliation widget)