[FIX] account_asset: multi-currency depreciation
- Create an asset category with a journal of type "Purchase" or
  "Miscellaneous".
- Create an asset in a currency different from the company currency
- Validate and post the first entry.

The error "The amount expressed in the secondary currency must be
positive..." is raised.

The sign is determined by the journal type. However, in case of an asset
depreciation or a deferred revenue, the amount is always positive. The
amount is negative in case of an asset created from a refund invoice
(see commit 4eab51f281d9).

There is actually no need of a `sign` variable. In case amount is
positive:
- first move is a credit, therefore `amount_currency` is negative
- second move is a debit, therefore `amount_currency` is positive

In case amoutn is negative:
- first move is a debit, therefore `amount_currency` is positive
- second move is a credit, therefore `amount_currency` is negative

opw-725358