[FIX] web: do not create new line if onchange has warning in o2m
In a one2many field, when the user clicks on the 'Add a line' button, we
need to create a new record.  This means that we call the default_get
route, then call the onchanges.  The onchange can return a 'warning'
key, which is displayed to the user.  But in the specific case of a
one2many, we don't want to actually create the record.  So, we reject the
deferred to prevent that from happening.