[FIX] base: field monetary lose context of t-lang
In monetary field refactoring (d0547901be) there was a small issue that
made context lost so the language use to render the field could be
different from the one to render the view.

For example: if the user has english, and print a report using the lang
of a partner in language spanish, the report would use spanish to format
number for all fields but monetary (which would use english).

opw-708197